Mopho 2: Realized that, when positioning GUI elements that have a predefined width and height, it's simpler to use setTopLeftPosition() rather than setBounds(), so I changed that throughout the project. Revised RandomizationOptions::getChildTreeForParam() to return a pointer instead of a reference. Realized that RandomizationOptions::getChildTreeForParam() isn't going to work for adding a listener to the tree - it returns a temporary copy of the tree, not the tree itself. Declared and defined addListenerToChildTreeForParam() and removeListenerFromChildTreeForParam(). Also declared and defined getCopyOfAllowedChoicesTreeForParam(), for use by the randomization methods. Finished defining GUI_Layer_AllowedChoices_OscShape.

+++ MophoGUI/MophoGUI 2.0/Source/randomization/0_rndm_comp_AllowChoiceToggles.h
@@ -14,7 +14,7 @@ class AllowChoiceToggles :
 	uint8 paramIndex;
 	UnexposedParameters* unexposedParams;
 	std::vector<std::unique_ptr<ToggleButton>> allowedChoiceToggles;
-	uint8 numberOfChoices;
+	int numberOfChoices;
 	int numberOfColumns;
 	int numberOfRows;
 	int toggle_w;
